rename useNameDates to useRegistrationDate
