+++ src/components/pages/profile/[name]/tabs/MoreTab/Miscellaneous.tsx
@@ -11,7 +11,7 @@ import FastForwardSVG from '@app/assets/FastForward.svg'
 import OutlinkSVG from '@app/assets/Outlink.svg'
 import { cacheableComponentStyles } from '@app/components/@atoms/CacheableComponent'
 import MobileFullWidth from '@app/components/@atoms/MobileFullWidth'
-import useRegistrationDate from '@app/hooks/useNameDates'
+import useRegistrationDate from '@app/hooks/useRegistrationDate'
 import { useSelfAbilities } from '@app/hooks/useSelfAbilities'
 import { useTransactionFlow } from '@app/transaction-flow/TransactionFlowProvider'
 import { formatDateTime, formatExpiry } from '@app/utils/utils'
